fix for ugly identifier renaming bug
problem seems to be that if you have code like:

{
	const xxx = "whatever";
	<T>xxx is {xxx}</T>
}
{
	const xxx = "whatever";
	<T>another proof that xxx is {xxx}</T>
}

then under the React Babel-preset (provided by create-react-app), the
identifier "xxx" gets renamed to "_xxx" in the second instance, probably in
order to resolve a name conflict (it doesn't happen if I manually rename "xxx"
to "xxx2" for instance). However, `traks update` uses a simpler Babel plugin
stack, and the identifier renaming doesn't happen there. This results in a
translation key mismatch because the dependency variable names are part of the
"signature" used to calculate the translation key. So you get a translation
that is not found - at runtime (`npm start`) it just shows up as an unknown
translation, but builds  (`TRAKS_BAKE_LANG=...  TRAKS_TRANSLATIONS_FILE=... npm
run build`) actually fail.

I should make a regression test convering this, but it's not completely trivial
to do...