solution for ReactDOMServer.renderToStaticMarkup()
Something changed between React 16.6.0 and 16.6.3 that broke code like this:

  ReactDOMServer.renderToStaticMarkup(<T>Translate me</T>)

Apparently, the above code no longer knows about the top-level <TraksProvider>
(a React context provider) that all <T>-tags must be a descendant of. Come to
think of it, I don't know how it _ever_ worked, because such <T>-tags were
_always_ inside a separate DOM tree :) So the React dev team probably fixed
some questionable code somewhere between those two releases.

Our project (aktiedysten.dk) using traks requires
ReactDOMServer.renderToStaticMarkup() style code in various places, for
instance, inside <svg> elements where <T> apparently doesn't work, and also in
a few hacks where we have to create DOM elements the old-school way
(document.createElement()).

The solution is an alternative to ReactDOMServer.renderToStaticMarkup() which
is provided by <TraksConsumer>; in addition to `lang` and `set_lang`, it now
also provides a `render_static` prop which is a function with the same call
semantics as ReactDOMServer.renderToStaticMarkup(), but which fixes the problem
by wrapping your element in an ad-hoc <TraksProvider>.

An added benefit of this is that components using `render_static` now
automatically update when the language changes; previously, code using
ReactDOMServer.renderToStaticMarkup() had a tendency to miss language changes,
and so you had to manually force another ReactDOMServer.renderToStaticMarkup()
call (e.g. via a re-render, or something else, depending on where it was
called) in order to see the language change in statically rendered markup.