Fix for TRAKS_FALLBACK_LANG bug
The export statement in the translations file is modified when baking a
language. This prevents wasting space on translations we're not using. Babel
can't really optimize this away because the data is in an object/dictionary, so
we're doing it manually. It seems I totally forgot about that in the previous
commit (f12d14feebfa60280c40f2b69c534ed8939a32a3); when setting
TRAKS_FALLBACK_LANG, we have a list of languages to try (`try_langs`)
containing two elements ([TRAKS_BAKE_LANG, TRAKS_FALLBACK_LANG]), but the code
that modifies the translations export was still only using TRAKS_BAKE_LANG, so
it would only contain TRAKS_BAKE_LANG translations. This causes setup-baked.js
to crash:

   opts.translations[key].apply(null, props.deps);

this causes an attempt to call apply() on undefined.